Fuse MapEntryAssertionSpecs (#622)
* Fuse MapAssertionSpecs
* Remove leftover from fusing IterableFeatureAssertionsSpec with
IterableAssertionSpec
* Comment-out IterableLike Scala specs, remove IterableFutureAssertionsSpec
