+++ misc/specs/atrium-specs-common/src/main/kotlin/ch/tutteli/atrium/specs/integration/MapEntryAssertionsSpec.kt
@@ -2,6 +2,7 @@ package ch.tutteli.atrium.specs.integration
 
 import ch.tutteli.atrium.api.fluent.en_GB.*
 import ch.tutteli.atrium.api.verbs.internal.expect
+import ch.tutteli.atrium.creating.Expect
 import ch.tutteli.atrium.specs.*
 import org.spekframework.spek2.Spek
 import org.spekframework.spek2.style.specification.Suite
@@ -9,6 +10,14 @@ import org.spekframework.spek2.style.specification.Suite
 abstract class MapEntryAssertionsSpec(
     isKeyValue: Fun2<Map.Entry<String, Int>, String, Int>,
     isKeyValueNullable: Fun2<Map.Entry<String?, Int?>, String?, Int?>,
+    keyFeature: Feature0<Map.Entry<String, Int>, String>,
+    key: Fun1<Map.Entry<String, Int>, Expect<String>.() -> Unit>,
+    valueFeature: Feature0<Map.Entry<String, Int>, Int>,
+    value: Fun1<Map.Entry<String, Int>, Expect<Int>.() -> Unit>,
+    nullableKeyFeature: Feature0<Map.Entry<String?, Int?>, String?>,
+    nullableKey: Fun1<Map.Entry<String?, Int?>, Expect<String?>.() -> Unit>,
+    nullableValueFeature: Feature0<Map.Entry<String?, Int?>, Int?>,
+    nullableValue: Fun1<Map.Entry<String?, Int?>, Expect<Int?>.() -> Unit>,
     describePrefix: String = "[Atrium] "
 ) : Spek({
 
@@ -16,15 +25,32 @@ abstract class MapEntryAssertionsSpec(
         describePrefix,
         isKeyValue.forSubjectLess("key", 1)
     ) {})
+
     include(object : SubjectLessSpec<Map.Entry<String?, Int?>>(
         "$describePrefix[nullable] ",
         isKeyValueNullable.forSubjectLess("key", 1)
     ) {})
 
+    include(object : KeyValueLikeAssertionsSpec<Map.Entry<String, Int>, Map.Entry<String?, Int?>>(
+        ::mapEntry,
+        ::mapEntry,
+        "key",
+        "value",
+        keyFeature,
+        key,
+        valueFeature,
+        value,
+        nullableKeyFeature,
+        nullableKey,
+        nullableValueFeature,
+        nullableValue,
+        describePrefix
+    ) {})
+
     fun describeFun(vararg pairs: SpecPair<*>, body: Suite.() -> Unit) =
         describeFunTemplate(describePrefix, pairs.map { it.name }.toTypedArray(), body = body)
 
-    val mapEntry = mapEntry("a", 1)
+    val mapEntry = mapEntry("hello", 1)
     val fluent = expect(mapEntry)
 
     describeFun(isKeyValue, isKeyValueNullable) {
@@ -32,13 +58,13 @@ abstract class MapEntryAssertionsSpec(
 
         context("map $mapEntry") {
             isKeyValueFunctions.forEach { (name, isKeyValueFun) ->
-                it("$name - a to 1 does not throw") {
-                    fluent.isKeyValueFun("a", 1)
+                it("$name - hello to 1 does not throw") {
+                    fluent.isKeyValueFun("hello", 1)
                 }
 
-                it("$name - a to 2 throws AssertionError") {
+                it("$name - hello to 2 throws AssertionError") {
                     expect {
-                        fluent.isKeyValueFun("a", 2)
+                        fluent.isKeyValueFun("hello", 2)
                     }.toThrow<AssertionError> {
                         message {
                             contains("value: 1", "$toBeDescr: 2")
@@ -51,7 +77,7 @@ abstract class MapEntryAssertionsSpec(
                         fluent.isKeyValueFun("b", 1)
                     }.toThrow<AssertionError> {
                         message {
-                            contains("key: \"a\"", "$toBeDescr: \"b\"")
+                            contains("key: \"hello\"", "$toBeDescr: \"b\"")
                             containsNot("value")
                         }
                     }
@@ -92,4 +118,5 @@ abstract class MapEntryAssertionsSpec(
             }
         }
     }
+
 })
